+++ NEWS.md
@@ -40,7 +40,7 @@ textclean 0.4.0 -
 **MINOR FEATURES**
 
 * `mgsub_regex` and `mgsub_fixed` to provide wrappers for `mgsub` that makes
-  their use apprent without setting the `fixed` command.
+  their use apparent without setting the `fixed` command.
   
 * `replace_curly_quote` added to replace curly quotes with straight versions.